[ build ] Major refactor of the build system.
Please find the detailed description of this refactor in the form of
an upcoming commit extending the documentation with a chapter on the
build system.

* Eliminates everything bootstrap related outside of the two
Makefiles.

* Adds a 'make release' target.

* Eliminates several sub makefiles from subdirs.

* Uses the $() convention in makefiles instead of ${}.

* Puts all build artifacts under build/ (except for the tests, because
for now we can't control how runtests invokes the idris binary).

* Changes the filename of intermediate outputs to include the CG's
name as file extension. This is needed for having better control in
the makefile to implement automated bootstrap shortcuts.

Discussion available at: https://github.com/idris-lang/Idris2/pull/1990